# frontend/openapi/context69.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "context69",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.8.1"
  },
  "paths": {
...
    "/v1/groups/by-path/{group_path}/documents/{document_id}/translations": {
      "get": {
        "tags": [],
        "operationId": "list_document_translation_jobs",
        "parameters": [
          {
            "name": "group_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRef"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/by-path/{group_path}/documents/{document_id}/translations/rebuild": {
      "post": {
        "tags": [],
        "operationId": "rebuild_document_translations",
        "parameters": [
          {
            "name": "group_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebuildDocumentTranslationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationJobsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/groups/by-path/{group_path}/translation-settings": {
      "get": {
        "tags": [],
        "operationId": "get_group_translation_settings",
        "parameters": [
          {
            "name": "group_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupTranslationSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "operationId": "update_group_translation_settings",
        "parameters": [
          {
            "name": "group_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupTranslationSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupTranslationSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/settings/translation": {
      "get": {
        "tags": [],
        "operationId": "get_translation_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSettingsResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "operationId": "update_translation_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslationSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSettingsResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/translation/providers": {
      "get": {
        "tags": [],
        "operationId": "list_translation_providers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationProviderPageResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "DeeplPlan": {
        "type": "string",
        "enum": [
          "free",
          "pro"
        ]
      },
...
      "GroupTranslationSettingsResponse": {
        "type": "object",
        "required": [
          "enabled",
          "default_target_locales",
          "glossary",
          "queued_count",
          "running_count",
          "succeeded_count",
          "failed_count"
        ],
        "properties": {
          "default_target_locales": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabled": {
            "type": "boolean"
          },
          "failed_count": {
            "type": "integer",
            "format": "int64"
          },
          "glossary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationGlossaryEntry"
            }
          },
          "queued_count": {
            "type": "integer",
            "format": "int64"
          },
          "running_count": {
            "type": "integer",
            "format": "int64"
          },
          "source_locale": {
            "type": [
              "string",
              "null"
            ]
          },
          "succeeded_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "RebuildDocumentTranslationsRequest": {
        "type": "object",
        "properties": {
          "target_locales": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TaskRef": {
        "type": "object",
        "required": [
          "task_id",
          "item_ids"
        ],
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "item_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "TranslationGlossaryEntry": {
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "properties": {
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        }
      },
      "TranslationJobResponse": {
        "type": "object",
        "required": [
          "job_id",
          "document_id",
          "target_locale",
          "status",
          "attempt_count",
          "source_character_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "attempt_count": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "document_id": {
            "type": "integer",
            "format": "int64"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "job_id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TranslationProviderKind"
              }
            ]
          },
          "source_character_count": {
            "type": "integer",
            "format": "int64"
          },
          "source_locale": {
            "type": [
              "string",
              "null"
            ]
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TranslationStatus"
          },
          "target_locale": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TranslationJobsResponse": {
        "type": "object",
        "required": [
          "jobs"
        ],
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationJobResponse"
            }
          }
        }
      },
      "TranslationLlmApiKind": {
        "type": "string",
        "enum": [
          "openai_responses",
          "openai_chat_completions",
          "anthropic_messages"
        ]
      },
      "TranslationProviderInput": {
        "type": "object",
        "required": [
          "provider",
          "priority"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "deepl_plan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeeplPlan"
              }
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "llm_api_kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TranslationLlmApiKind"
              }
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "monthly_character_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "provider": {
            "$ref": "#/components/schemas/TranslationProviderKind"
          }
        }
      },
      "TranslationProviderKind": {
        "type": "string",
        "enum": [
          "deepl",
          "llm",
          "libretranslate"
        ]
      },
...
      "TranslationProviderPageResponse": {
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationProviderResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "TranslationProviderResponse": {
        "type": "object",
        "required": [
          "provider",
          "enabled",
          "priority",
          "has_api_key",
          "current_month_characters"
        ],
        "properties": {
          "current_month_characters": {
            "type": "integer",
            "format": "int64"
          },
          "deepl_plan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeeplPlan"
              }
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "has_api_key": {
            "type": "boolean"
          },
          "llm_api_kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TranslationLlmApiKind"
              }
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "monthly_character_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "provider": {
            "$ref": "#/components/schemas/TranslationProviderKind"
          }
        }
      },
      "TranslationSettingsResponse": {
        "type": "object",
        "required": [
          "providers"
        ],
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationProviderResponse"
            }
          }
        }
      },
      "TranslationStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "skipped",
          "quota_exceeded",
          "unavailable"
        ]
      },
...
      "UpdateGroupTranslationSettingsRequest": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "default_target_locales": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabled": {
            "type": "boolean"
          },
          "glossary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationGlossaryEntry"
            }
          },
          "source_locale": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateTranslationSettingsRequest": {
        "type": "object",
        "required": [
          "providers"
        ],
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationProviderInput"
            }
          }
        }
      },
...
    }
  },
...
}